Create i2p.noscript.config.secured.txt
Initial commit of safer NoScript configuration. My assumption is that this is leftover doodoo from TorBrowser's insane implementation and configuration of NoScript before they started slowly changing it to be less bad. Either that, or the same person who is responsible for the TorBrowser bugdoor is working on I2P Browser as well. Fun times!